requirements.txt
# Core dependencies
python-markdown>=3.4.0
pyyaml>=6.0
click>=8.0.0

# AI/ML dependencies
openai>=1.0.0
tiktoken>=0.5.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Optional dependencies for online repository support
GitPython>=3.1.0  # For git cloning
PyGithub>=1.59.0  # For GitHub API access
requests>=2.31.0  # For HTTP requests

# Optional dependencies
rich>=13.0.0  # For better CLI output